requirements.txt
# Core Dependencies
apache-airflow>=2.5.0
google-cloud-storage>=2.10.0
google-cloud-dataproc>=3.0.0
pyspark>=3.3.0

# Image Processing
Pillow>=9.5.0
opencv-python>=4.7.0
numpy>=1.24.0

# Machine Learning
tensorflow>=2.12.0
pandas>=2.0.0
pyarrow>=12.0.0

# Data Versioning
dvc[gcs]>=3.0.0
dvc-s3>=2.23.0  # Alternative if using S3

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Utilities
tqdm>=4.65.0
google-auth>=2.17.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0

# Development (optional)
pytest>=7.3.0
black>=23.3.0
flake8>=6.0.0

